+++ src/main/java/com/project/fintech/application/AuthApplication.java
@@ -4,7 +4,7 @@ import com.project.fintech.auth.constants.RedisKeyConstants;
 import com.project.fintech.model.dto.ResponseDto;
 import com.project.fintech.auth.jwt.JwtUtil;
 import com.project.fintech.auth.otp.OtpUtil;
-import com.project.fintech.auth.otp.OtpVerificationDto;
+import com.project.fintech.model.dto.OtpVerificationDto;
 import com.project.fintech.service.AuthService;
 import com.project.fintech.exception.CustomException;
 import com.project.fintech.exception.ErrorCode;
@@ -17,7 +17,6 @@ import jakarta.servlet.http.HttpServletResponse;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
