+++ src/main/java/com/project/fintech/model/dto/UserEmailDto.java
@@ -1,6 +1,5 @@
 package com.project.fintech.model.dto;
 
-import jakarta.validation.Valid;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Builder;
@@ -8,7 +7,7 @@ import lombok.Data;
 
 @Data
 @Builder
-public class EmailVerificationDto {
+public class UserEmailDto {
     @Email(message = "이메일 형식이 유효하지 않습니다.")
     @NotBlank(message = "이메일 입력은 필수 입니다.")
     private String email;
